fix: Update failed test case for 0-5A wallet switching
